+++ monstersWeapons/Weapon.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
 class Weapon {
-    int damage;
-    int durability;
+    short damage;
+    short durability;
 
 public:
-    Weapon(const int& dmg, const int& dur) : damage(dmg), durability(dur) {}
-    int getDamage() const { return damage; }
-    int getDurability() const { return durability; }
-    int reduceDurability(const int& dmg) { return durability -= dmg; }
+    Weapon(const short& dmg, const short& dur) : damage(dmg), durability(dur) {}
+    short getDamage() const { return damage; }
+    short getDurability() const { return durability; }
+    short reduceDurability(const short& dmg) { return durability -= dmg; }
 };
